header organizing, collision detector prototypes, collision functions to t_object struct
