+++ inc/object.h
@@ -3,23 +3,15 @@
 
 # include "../inc/vector.h"
 
-typedef struct	s_sphere
-{
-	t_vector	center;
-	double		diameter;
-	int			color;
-}	t_sphere;
-
-/*				ambient	camera	light	sphere	plane	cylinder
-   type			YES		YES		YES		YES		YES		YES
-   color		YES		NO		NO		YES		YES		YES
+/* type			ambient	camera	light	sphere	plane	cylinder
+   color		YES		NO		NO*		YES		YES		YES
    brightness	YES		NO		YES		NO		NO		NO
    location		NO		YES		YES		YES		YES		YES
    orientation	NO		YES		NO		NO		YES		YES
    fov			NO		YES		NO		NO		NO		NO
    diameter		NO		NO		NO		YES		NO		YES
    height		NO		NO		NO		NO		NO		YES
-   */
+   *) light has color in bonus */
 typedef struct	s_object
 {
 	t_type		type;
@@ -30,6 +22,7 @@ typedef struct	s_object
 	int			fov;
 	double		diameter;
 	double		height;
+	int			collision;
 }	t_object;
 
 typedef enum e_type
@@ -43,7 +36,25 @@ typedef enum e_type
 }	t_type;
 
 t_object	*parse_object(char *line);
-t_sphere	*create_sphere(t_vector center, double diameter, int color);
-int			sphere_collision(t_vector d, t_sphere *sphere, double *distance);
+
+/* Collision detectors 
+ * 	Parameters:
+ * 		-ray as a direction vector (normalized? might not matter)
+ * 		-shape itself
+ * 		-distance pointer
+ * 	Returns: color of visible point (final color?) */
+int			sphere_collision(t_vector d, t_object *sphere, double *distance);
+int			plane_collision(t_vector d, t_object *plane, double *distance);
+int			cylinder_collision(t_vector d, t_object *cylinder, double *distance);
+
+/* Shaped object creators */
+t_object	*create_sphere(char **info);
+t_object	*create_plane(char **info);
+t_object	*create_cylinder(char **info);
+
+/* Non-shaped object creators */
+t_object	*create_ambient(char **info);
+t_object	*create_camera(char **info);
+t_object	*create_light(char **info);
 
 #endif
